feat: Add a hacky loadQuery method to call exec on the driver
This makes all the multiple statement queries in the postgres tests pass without modification.

Can we make it possible call methods on the driver directly? Kysely doesn't seem to offer this functionality on the
connection. It only has executeQuery and streamQuery in the interface and that is not enough for this.